moved hashlock_result to Raiden
